remove NativeDatePickerCore.tsx
