+++ apps/fluent-tester/src/FluentTester.tsx
@@ -53,8 +53,11 @@ const TestListSeparator = Separator.customize((t) => ({
 }));
 
 export const FluentTester: React.FunctionComponent<FluentTesterProps> = (props: FluentTesterProps) => {
+  // filters for tests specific to each platform
+  const filteredTestComponents = tests.filter((test) => test.platforms.includes(Platform.OS as string));
+
   // sort tests alphabetically by name
-  const sortedTestComponents = tests.sort((a, b) => a.name.localeCompare(b.name));
+  const sortedTestComponents = filteredTestComponents.sort((a, b) => a.name.localeCompare(b.name));
 
   const { initialTest, enableSinglePaneView } = props;
   const initialSelectedTestIndex = sortedTestComponents.findIndex((description) => {
